(initram) encryption-helper: No send the signal when userdata is full encrypted
Because the full userdata luks implementation, we want to avoid
sending a signal to the service.

This encryption mode is unattended and is configured by the
recovery installer zip, not finished yet.

The method used to disable the signal, is the userdata_luks var implementation.
When the value is "full", the function signal_root_move_done returns.

Signed-off-by: berbascum <berbascum@ticv.cat>